requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.9.14=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
contourpy=1.0.5=pypi_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
docker-pycreds=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.37.4=pypi_0
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20220512=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
jpeg=9e=h7f8727e_0
kaggle=1.5.12=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.6.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
portalocker=2.5.1=pypi_0
promise=2.3=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.2=pypi_0
ptflops=0.6.9=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python-slugify=6.1.2=pypi_0
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py38h06a4308_0
sentry-sdk=1.9.9=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=63.4.1=py38h06a4308_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.39.3=h5082296_0
tabulate=0.8.10=pypi_0
termcolor=2.0.1=pypi_0
text-unidecode=1.3=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py38_cu113
torchvision=0.13.1=py38_cu113
tqdm=4.64.1=pypi_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.11=py38h06a4308_0
wandb=0.13.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yacs=0.1.8=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
